+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [0.27.18] - 2021-12-04
+## [0.27.19] - 2021-12-05
 ### Changed
  - mimixbox command.
    - When an error occurs, the name of the applet that caused the error is displayed.
@@ -20,6 +20,9 @@ All notable changes to this project will be documented in this file. The format 
    - Print error if the -r option is not attached when copying the directory.
    - Print error if the copy destination is in the hierarchy below the copy source directory.
    - Fixed the bug that the directory hierarchy of the source path is copied as it is when copying the directory. The correct process is to copy from the base name of the source path.
+ - md5sum / sha1sum / sha256sum / sha512sum command.
+   - Print the error if the argument is directory or non-existent file.
+   - If PIPE and file path are passed at the same time, the PIPE data will be ignored.
 ## [0.27.10] - 2021-12-01
 ### Added
  - Add ShellSpec tetsing framework for integration test.